Moved ctid cursor test to its own file CtidEnabledPostgresSourceTest
