+++ airbyte-integrations/connectors/source-jdbc/src/testFixtures/java/io/airbyte/integrations/source/jdbc/test/JdbcSourceAcceptanceTest.java
@@ -757,7 +757,9 @@ public abstract class JdbcSourceAcceptanceTest {
       airbyteStream.setCursorField(List.of(COL_ID));
       airbyteStream.setDestinationSyncMode(DestinationSyncMode.APPEND);
     });
-    final List<AirbyteMessage> actualMessagesFirstSync = readCatalogForInitialSync(configuredCatalog);
+
+    final List<AirbyteMessage> actualMessagesFirstSync = MoreIterators
+        .toList(source.read(config, configuredCatalog, createEmptyState(streamName, namespace)));
 
     // get last state message.
     final Optional<AirbyteMessage> stateAfterFirstSyncOptional = actualMessagesFirstSync.stream()
@@ -771,13 +773,31 @@ public abstract class JdbcSourceAcceptanceTest {
 
     // Represents the state after the first stream has been updated
     final List<DbStreamState> expectedStateStreams1 = List.of(
-        buildExpectedStreamState(streamName, namespace, List.of(COL_ID), "3", 1L),
-        buildExpectedStreamState(streamName2, namespace, List.of(COL_ID), null, null));
+        new DbStreamState()
+            .withStreamName(streamName)
+            .withStreamNamespace(namespace)
+            .withCursorField(List.of(COL_ID))
+            .withCursor("3")
+            .withCursorRecordCount(1L),
+        new DbStreamState()
+            .withStreamName(streamName2)
+            .withStreamNamespace(namespace)
+            .withCursorField(List.of(COL_ID)));
 
     // Represents the state after both streams have been updated
     final List<DbStreamState> expectedStateStreams2 = List.of(
-        buildExpectedStreamState(streamName, namespace, List.of(COL_ID), "3", 1L),
-        buildExpectedStreamState(streamName2, namespace, List.of(COL_ID), "3", 1L));
+        new DbStreamState()
+            .withStreamName(streamName)
+            .withStreamNamespace(namespace)
+            .withCursorField(List.of(COL_ID))
+            .withCursor("3")
+            .withCursorRecordCount(1L),
+        new DbStreamState()
+            .withStreamName(streamName2)
+            .withStreamNamespace(namespace)
+            .withCursorField(List.of(COL_ID))
+            .withCursor("3")
+            .withCursorRecordCount(1L));
 
     final List<AirbyteMessage> expectedMessagesFirstSync = new ArrayList<>(getTestMessages());
     expectedMessagesFirstSync.add(createStateMessage(expectedStateStreams1.get(0), expectedStateStreams1));
@@ -854,7 +874,6 @@ public abstract class JdbcSourceAcceptanceTest {
     assertTrue(firstSyncState.get("cursor").asText().contains("2021-01-01"));
     assertTrue(firstSyncState.get("cursor").asText().contains("00:00:00"));
     assertEquals(2L, firstSyncState.get("cursor_record_count").asLong());
-    addStandardStateTypeToSyncState(firstSyncState);
 
     final List<String> firstSyncNames = firstSyncActualMessages.stream()
         .filter(r -> r.getType() == Type.RECORD)
@@ -883,7 +902,6 @@ public abstract class JdbcSourceAcceptanceTest {
     assertTrue(secondSyncState.get("cursor").asText().contains("2021-01-02"));
     assertTrue(secondSyncState.get("cursor").asText().contains("00:00:00"));
     assertEquals(1L, secondSyncState.get("cursor_record_count").asLong());
-    addStandardStateTypeToSyncState(secondSyncState);
 
     final List<String> secondSyncNames = secondSyncActualMessages.stream()
         .filter(r -> r.getType() == Type.RECORD)
@@ -935,12 +953,12 @@ public abstract class JdbcSourceAcceptanceTest {
     throw new IllegalArgumentException("Stream not found in state message: " + streamName);
   }
 
-  protected void incrementalCursorCheck(
-                                        final String initialCursorField,
-                                        final String cursorField,
-                                        final String initialCursorValue,
-                                        final String endCursorValue,
-                                        final List<AirbyteMessage> expectedRecordMessages)
+  private void incrementalCursorCheck(
+                                      final String initialCursorField,
+                                      final String cursorField,
+                                      final String initialCursorValue,
+                                      final String endCursorValue,
+                                      final List<AirbyteMessage> expectedRecordMessages)
       throws Exception {
     incrementalCursorCheck(initialCursorField, cursorField, initialCursorValue, endCursorValue,
         expectedRecordMessages,
@@ -962,31 +980,32 @@ public abstract class JdbcSourceAcceptanceTest {
     final ConfiguredAirbyteCatalog configuredCatalog = new ConfiguredAirbyteCatalog()
         .withStreams(List.of(airbyteStream));
 
-    final DbStreamState dbStreamState = new DbStreamState()
-        .withStreamName(airbyteStream.getStream().getName())
-        .withStreamNamespace(airbyteStream.getStream().getNamespace())
-        .withCursorField(List.of(initialCursorField))
-        .withCursor(initialCursorValue)
-        .withCursorRecordCount(1L);
+    final DbStreamState dbStreamState = buildStreamState(airbyteStream, initialCursorField, initialCursorValue);
 
     final List<AirbyteMessage> actualMessages = MoreIterators
         .toList(source.read(config, configuredCatalog, Jsons.jsonNode(createState(List.of(dbStreamState)))));
 
     setEmittedAtToNull(actualMessages);
 
-    final List<DbStreamState> expectedStreams = List.of(
-        new DbStreamState()
-            .withStreamName(airbyteStream.getStream().getName())
-            .withStreamNamespace(airbyteStream.getStream().getNamespace())
-            .withCursorField(List.of(cursorField))
-            .withCursor(endCursorValue)
-            .withCursorRecordCount(1L));
+    final List<DbStreamState> expectedStreams = List.of(buildStreamState(airbyteStream, cursorField, endCursorValue));
+
     final List<AirbyteMessage> expectedMessages = new ArrayList<>(expectedRecordMessages);
     expectedMessages.addAll(createExpectedTestMessages(expectedStreams));
 
     assertEquals(expectedMessages.size(), actualMessages.size());
     assertTrue(expectedMessages.containsAll(actualMessages));
     assertTrue(actualMessages.containsAll(expectedMessages));
+  }
+
+  protected DbStreamState buildStreamState(final ConfiguredAirbyteStream configuredAirbyteStream,
+                                           final String cursorField,
+                                           final String cursorValue) {
+    return new DbStreamState()
+        .withStreamName(configuredAirbyteStream.getStream().getName())
+        .withStreamNamespace(configuredAirbyteStream.getStream().getNamespace())
+        .withCursorField(List.of(cursorField))
+        .withCursor(cursorValue)
+        .withCursorRecordCount(1L);
   }
 
   // get catalog and perform a defensive copy.
@@ -998,6 +1017,7 @@ public abstract class JdbcSourceAcceptanceTest {
             .collect(Collectors.toList()));
     return catalog;
   }
+
   protected AirbyteCatalog getCatalog(final String defaultNamespace) {
     return new AirbyteCatalog().withStreams(List.of(
         CatalogHelpers.createAirbyteStream(
@@ -1252,49 +1272,6 @@ public abstract class JdbcSourceAcceptanceTest {
       return new AirbyteMessage().withType(Type.STATE).withState(new AirbyteStateMessage().withType(AirbyteStateType.LEGACY)
           .withData(Jsons.jsonNode(new DbState().withCdc(false).withStreams(legacyStates))));
     }
-  }
-
-  /**
-   * Method to add state_type: "standard" to state message for CTID based syncs
-   *
-   * @param syncState
-   */
-  protected void addStandardStateTypeToSyncState(final JsonNode syncState) {
-    // no op
-  }
-
-  /**
-   * Read catalog to build iterators for the first sync
-   *
-   * @param configuredCatalog catalog of DB source
-   * @return A list of Airbyte messages generated during sync
-   * @throws Exception
-   */
-  protected List<AirbyteMessage> readCatalogForInitialSync(final ConfiguredAirbyteCatalog configuredCatalog) throws Exception {
-    return MoreIterators
-        .toList(source.read(config, configuredCatalog, createEmptyState(streamName, getDefaultNamespace())));
-  }
-
-  /**
-   * Build the expectged DBStreamState object
-   */
-  protected DbStreamState buildExpectedStreamState(final String streamName,
-                                                   final String nameSpace,
-                                                   final List<String> cursorFields,
-                                                   final String cursorValue,
-                                                   final Long cursorRecordCount) {
-    final DbStreamState streamState = new DbStreamState();
-    streamState.setStreamName(streamName);
-    streamState.setStreamNamespace(nameSpace);
-    streamState.setCursorField(cursorFields);
-    if (cursorValue != null) {
-      streamState.setCursor(cursorValue);
-    }
-
-    if (cursorRecordCount != null) {
-      streamState.setCursorRecordCount(cursorRecordCount);
-    }
-    return streamState;
   }
 
 }
